+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2026-07-08:
+  - updated packages.
+  - fix to `UdpTextFilter.ApplyAsync` which did not pass `TagListEntry` objects to `UdpChunkBuilder`.
+
 ### [12.0.2] - 2026-06-27
 
 - 2026-06-27: updated packages.